moved exceptions to new folder
