+++ lib/model/storage/exceptions/bad_certificate_exception.dart
