+++ lib/widgets/home_paths/home.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:taskwarrior/model/storage/bad_certificate_exception.dart';
+import 'package:taskwarrior/model/storage/exceptions/bad_certificate_exception.dart';
 import 'package:taskwarrior/widgets/home_paths.dart';
 import 'package:taskwarrior/widgets/taskserver.dart' as rc;
